requirements.txt
pip==23.3
setuptools
torch
torchvision
torchaudio --index-url https://download.pytorch.org/whl/cu124
omegaconf
cloudpickle
fvcore
ftfy
tqdm
git+https://github.com/openai/CLIP.git
timm
transformers
datasets
black
ipython
ipykernel
ultralytics
more_itertools
wandb
accelerate
simple_parsing
natsort
einops
scipy
opencv-python
hydra-core
SciencePlots
google-genai
openai
shapely
scikit-learn
hdbscan
umap-learn
uuid
plotly
polars